fix: Rename sshlistners package to sshlisteners
Closes #48
